Fixing chat crash on macOS (and likely Linux)
Long is 8 bytes outside of Windows, not 4. Hardcoded uint32_t is causing pointer corruption on the stack because Python expects an 8 byte type.

+++ Sources/Plasma/FeatureLib/pfPython/cyMiscGlue.cpp
@@ -232,7 +232,7 @@ PYTHON_GLOBAL_METHOD_DEFINITION(PtSendRTChat, args, "Params: fromPlayer,toPlayer
     PyObject* fromPlayerObj = nullptr;
     PyObject* toPlayerListObj = nullptr;
     ST::string message;
-    uint32_t msgFlags = 0;
+    long msgFlags = 0;
     const char* err = "PtSendRTChat expects a ptPlayer, a sequence of ptPlayers, a string, and an optional long";
 
     if (!PyArg_ParseTuple(args, "OOO&|l", &fromPlayerObj, &toPlayerListObj,
